+++ site-cookbooks/pxe-server/metadata.rb
@@ -1,7 +1,7 @@
 name             'pxe-server'
-maintainer       'YOUR_COMPANY_NAME'
-maintainer_email 'YOUR_EMAIL'
-license          'All rights reserved'
+maintainer       'sanritz automation'
+maintainer_email 'kitsunai@sanritz.co.jp'
+license          'MIT'
 description      'Installs/Configures pxe-server'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.1.0'